can upload and download csv price file

+++ indicator/ma_graph.py
@@ -21,7 +21,7 @@ def bolband(price, days, factor):
     sma = SMA(price, days)
     std = price.std()
     # factor = 1
-    price['uband'] = pd.DataFrame(sma + int(factor)*std)
-    price['lband'] = pd.DataFrame(sma - int(factor)*std)
+    price['uband'] = pd.DataFrame(sma + float(factor)*std)
+    price['lband'] = pd.DataFrame(sma - float(factor)*std)
 
     return price
